requirements.txt
asgiref==3.11.0
babel==2.17.0
certifi==2026.2.25
charset-normalizer==3.4.4
customtkinter==5.2.2
darkdetect==0.8.0
Django==5.2.8
django-cors-headers==4.9.0
django-crontab==0.7.1
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
idna==3.11
packaging==25.0
pdfkit==1.0.0
Pillow==9.5.0
PyJWT==2.12.1
reportlab==4.4.4
requests==2.33.1
sqlparse==0.5.3
tkcalendar==1.6.1
tzdata==2025.2
urllib3==2.6.3
gunicorn
whitenoise
dj-database-url
psycopg[binary]